+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/utils/LoggingUtils.kt
@@ -6,14 +6,14 @@ import io.ktor.util.logging.Logger
 internal expect inline fun getLogger(lvl: LogLvl, tag: String): Logger
 
 @Suppress("NOTHING_TO_INLINE")
-inline fun Logger.info(message: () -> String) = info(message())
+internal inline fun Logger.info(message: () -> String) = info(message())
 
 @Suppress("NOTHING_TO_INLINE")
-inline fun Logger.warn(message: () -> String) = warn(message())
+internal inline fun Logger.warn(message: () -> String) = warn(message())
 
 @Suppress("NOTHING_TO_INLINE")
-inline fun Logger.debug(message: () -> String) = debug(message())
+internal inline fun Logger.debug(message: () -> String) = debug(message())
 
 @Suppress("NOTHING_TO_INLINE")
-inline fun Logger.error(throwable: Throwable? = null, message: () -> String) =
+internal inline fun Logger.error(throwable: Throwable? = null, message: () -> String) =
     throwable?.let { error(message(), it) } ?: error(message())